requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==3.0.5
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
cron-descriptor==1.4.0
cryptography==41.0.7
daphne==3.0.2
distlib==0.3.8
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.11
django-appconf==1.0.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-compressor==4.3.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.11.2
django-guardian==2.4.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
faiss-cpu==1.8.0
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gptcache==0.1.43
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jsonfield==3.1.0
kombu==5.3.4
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
mysqlclient==2.2.1
numpy==1.26.4
onnxruntime==1.14.1
openai==1.25.0
packaging==24.0
phonenumbers==8.13.27
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.1
pydantic-core==2.18.2
pyOpenSSL==23.3.0
pyreadline3==3.4.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-ulid==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
redis==5.0.4
redis-om==0.3.1
regex==2024.4.28
reportlab==4.0.8
requests==2.31.0
rjsmin==1.2.1
safetensors==0.4.3
service-identity==23.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
supervisor==4.2.5
sympy==1.12
tiktoken==0.6.0
tokenizers==0.19.1
tqdm==4.66.2
transformers==4.40.1
twisted==23.10.0
txaio==23.1.1
types-cffi==1.16.0.20240331
types-pyOpenSSL==24.1.0.20240425
types-redis==4.6.0.20240425
types-setuptools==69.5.0.20240513
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
yarl==1.9.4
zope.interface==6.1
